requirements.txt
python-telegram-bot
nsetools
os